website/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
Django==5.1.7
django-cors-headers==4.7.0
django-filter==25.1
django-livereload-server==0.5.1
djangorestframework==3.16.0
et_xmlfile==2.0.0
fonttools==4.56.0
idna==3.10
joblib==1.4.2
kiwisolver==1.4.8
Markdown==3.7
matplotlib==3.10.1
narwhals==1.30.0
numpy==2.2.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
patsy==1.0.1
pillow==11.1.0
plotly==6.0.0
plotly-express==0.4.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
statsmodels==0.14.4
threadpoolctl==3.6.0
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
